+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- Add `Pages.handle_redirect/1` and callback in `Pages.Driver.LiveView`, for cases when `handle_info/2` may issue a
+  redirect to the client.
 - **Breaking change:** Drop support for Elixir 1.13.
 
 ## 0.13.4